+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.13.5] - 2026-01-23
+
+### Changed
+- Ensure GitHub auth scopes/permissions and improve Firebase key fetch fallback
+- Always set Firebase database URL in .env during setup
+
 ## [2.13.4] - 2026-01-23
 
 ### Changed